Extract sync cutover helpers
